functions/consume_invoices_func/e2e-test/requirements.txt
google-cloud-storage==1.23.0
google-cloud-logging==1.14.0
